requirements.txt
asgiref==3.9.1
colorama==0.4.6
dj-database-url==3.0.1
Django==5.2.4
django-filter==25.1
djangorestframework==3.16.0
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.1
pytest-django==4.11.1
python-decouple==3.8
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
